Safari-609.1.11.1/LayoutTests/http/tests/security/window-named-valueOf-expected.txt

CONSOLE MESSAGE: line 1: Threw exception: TypeError: Cannot set prototype of immutable prototype object
This passes if it doesn't alert the contents of innocent-victim.  
